setvalues S3 class

Functions for specifying set values to be input into the values field of set(). It also has 3 child classes, "finite", "interval" and "sequence".

Usage

setvalues_finite(values)

setvalues_interval(from, to)

setvalues_sequence(sequence_function, initial_value)

Arguments

values

numeric
Vector of values representing a finite set

from

numeric

Lower bound of the interval.

to

numeric
Upper bound of the interval.

sequence_function

function
A function with one argument (representing the previous value) for constructing the sequence.

initial_value

function
Initial value of the sequence.

Value

An object of (s3) class c(type, "setvalues"), where type is one of "finite", "interval", or "sequence".

Details

The following three types of set values are available:

setvalues_finite() is for finite sets for example: {0,1}{0, 1}{0,1}

setvalues_interval() for continuous intervals of the real domain, for example: [0,8)[0,\infty)[0,8)

setvalues_sequences() sequences with countably infinite cardinality, defined as recurrence relations such that: an+1=f(an)愼㸰with愼㸰known愼㸰inital愼㸰value愼㸰a0a_{n+1} = f(a_n) \text{ with known inital value } a_0an+1=f(an)愼㸰with愼㸰known愼㸰inital愼㸰value愼㸰a0 for example: N愼㸰can愼㸰be愼㸰constructed愼㸰with愼㸰f(x)=x+1愼㸰and愼㸰a0=0\mathbb N \text{ can be constructed with } f(x) = x + 1 \text{ and } a_0 = 0N愼㸰can愼㸰be愼㸰constructed愼㸰with愼㸰f(x)=x+1愼㸰and愼㸰a0=0

Note: These are not exhaustive but are sufficient for the purposes of this package (i.e. simple sets to represent the support or allowable paramter values of a distribution).

Examples

setvalues_finite(c(0, 1))
setvalues_interval(0, Inf)
setvalues_sequence(function(x) x + 1, 1)